% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prakriti-package.R
\docType{package}
\name{prakriti-package}
\alias{prakriti}
\alias{prakriti-package}
\title{prakriti: Color palettes inspired by India's natural landscapes}
\description{
30 hand-tuned color palettes drawn from India's biomes and geography -
Himalayan snow, Thar dunes, Kerala backwaters, Andaman reefs, Spiti's cold
desert, Kashmir's autumn chinar, and more. Includes ggplot2 scales,
continuous interpolation, and base-graphics display helpers.
}
\section{Main functions}{

\itemize{
\item \code{\link[=prakriti_palette]{prakriti_palette()}} - return a vector of hex codes
\item \code{\link[=scale_color_prakriti]{scale_color_prakriti()}} / \code{\link[=scale_fill_prakriti]{scale_fill_prakriti()}} - ggplot2 scales
\item \code{\link[=display_prakriti]{display_prakriti()}} - plot palettes
\item \code{\link[=prakriti_names]{prakriti_names()}} / \code{\link[=prakriti_info]{prakriti_info()}} - discover what's available
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/orijitghosh/prakriti}
  \item Report bugs at \url{https://github.com/orijitghosh/prakriti/issues}
}

}
\author{
\strong{Maintainer}: Arijit Ghosh \email{arijitghosh2009@gmail.com}

}
\keyword{internal}
