% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prakriti_names}
\alias{prakriti_names}
\title{Names of available prakriti palettes}
\usage{
prakriti_names()
}
\value{
A character vector of palette names.
}
\description{
Names of available prakriti palettes
}
\examples{
prakriti_names()
}
